refactor: move shared types & function to handler
